+++ CHANGES.md
@@ -37,6 +37,7 @@
 - #1445: Participants list uses big font in embedded mode
 - #1455: Avatar in controlbox status-view not updated
 - #1465: When highlighting a roster contact, they're incorrectly shown as online
+- #1494: Trim whitespace around messages
 - #1495: Mentions should always include a URI attribute
 - #1502: Fatal error when using prebind
 - #1532: Converse reloads on enter pressed in the filter box